==> notebooks/1_get_data_nb.ipynb <==
import pandas as pd
import sys
import numpy as np
import os

sys.path.append('..')
from utils import get_historical_data_wave, get_historical_data_tide, get_historical_data_wind,test_api
sys.path.pop()

test_api()

wind_Kn_hist = get_historical_data_wind(location='Knokke')
wind_Dp_hist = get_historical_data_wind(location='Nieuwpoort')
wind_Oo_hist = get_historical_data_wind(location='Oostende')

wave_Kn_hist = get_historical_data_wave(2000,2023,'ZHG')
wave_Dp_hist = get_historical_data_wave(2000,2023,'NPB')
wave_Oo_hist = get_historical_data_wave(2000,2023,'OST')

tide_hist = get_historical_data_tide()

wave_Kn_hist

combined_Kn = wave_Kn_hist.merge(wind_Kn_hist, on='Timestamp', how='left')
combined_Kn = combined_Kn.merge(tide_hist, on='Timestamp', how='left')

combined_Dp = wave_Dp_hist.merge(wind_Dp_hist, on='Timestamp', how='left')
combined_Dp = combined_Dp.merge(tide_hist, on='Timestamp', how='left')

combined_Oo = wave_Oo_hist.merge(wind_Oo_hist, on='Timestamp', how='left')
combined_Oo = combined_Oo.merge(tide_hist, on='Timestamp', how='left')

combined_Kn = combined_Kn.dropna(axis = 0, how = 'all')
combined_Dp = combined_Dp.dropna(axis = 0, how = 'all')
combined_Oo = combined_Oo.dropna(axis = 0, how = 'all')

combined_Kn.to_csv('../raw_data/Kn_data.csv')
combined_Dp.to_csv('../raw_data/Dp_data.csv')
combined_Oo.to_csv('../raw_data/Oo_data.csv')
